Moved VideoCell to separate file
